refactor(serializer): Move DUMP logic into RdbSerializer
